refactor: Remove OnboardingViewModel
