+++ Mail/Views/Onboarding/OnboardingView.swift
@@ -23,17 +23,55 @@ import MailCore
 import MailResources
 import SwiftUI
 
+struct Slide: Identifiable {
+    let id: Int
+    let backgroundImage: Image
+    let animationFile: String
+    let title: String
+    let description: String
+
+    static let allSlides = [
+        Slide(
+            id: 1,
+            backgroundImage: Image(resource: MailResourcesAsset.onboardingBackground1),
+            animationFile: "illu_1",
+            title: MailResourcesStrings.Localizable.onBoardingTitle1,
+            description: ""
+        ),
+        Slide(
+            id: 2,
+            backgroundImage: Image(resource: MailResourcesAsset.onboardingBackground2),
+            animationFile: "illu_2",
+            title: MailResourcesStrings.Localizable.onBoardingTitle2,
+            description: MailResourcesStrings.Localizable.onBoardingDescription2
+        ),
+        Slide(
+            id: 3,
+            backgroundImage: Image(resource: MailResourcesAsset.onboardingBackground3),
+            animationFile: "illu_3",
+            title: MailResourcesStrings.Localizable.onBoardingTitle3,
+            description: MailResourcesStrings.Localizable.onBoardingDescription3
+        ),
+        Slide(
+            id: 4,
+            backgroundImage: Image(resource: MailResourcesAsset.onboardingBackground4),
+            animationFile: "illu_4",
+            title: MailResourcesStrings.Localizable.onBoardingTitle4,
+            description: MailResourcesStrings.Localizable.onBoardingDescription4
+        )
+    ]
+}
+
 struct OnboardingView: View {
-    @StateObject var viewModel = OnboardingViewModel()
+    @Environment(\.window) private var window
+    @Environment(\.dismiss) private var dismiss
 
     @State private var selection: Int
     @State private var presentAlert = false
     @State private var isLoading = false
 
     private var isScrollEnabled: Bool
-
-    @Environment(\.window) var window
-    @Environment(\.dismiss) private var dismiss
+    private var slides = Slide.allSlides
 
     init(page: Int = 1, isScrollEnabled: Bool = true) {
         _selection = State(initialValue: page)
@@ -45,11 +83,11 @@ struct OnboardingView: View {
     var body: some View {
         VStack(spacing: 0) {
             Group {
-                if !isScrollEnabled, let slide = viewModel.slides.first { $0.id == selection } {
+                if !isScrollEnabled, let slide = slides.first { $0.id == selection } {
                     SlideView(slide: slide)
                 } else {
                     TabView(selection: $selection) {
-                        ForEach(viewModel.slides) { slide in
+                        ForEach(slides) { slide in
                             SlideView(slide: slide)
                                 .tag(slide.id)
                         }
@@ -63,12 +101,12 @@ struct OnboardingView: View {
                     .resizable()
                     .scaledToFit()
                     .frame(height: Constants.onboardingLogoHeight)
-                    .padding(.top, 16)
+                    .padding(.top, 28)
             }
 
             // Buttons
             VStack(spacing: 24) {
-                if selection == viewModel.slides.count {
+                if selection == slides.count {
                     // Show login button
                     LargeButton(title: MailResourcesStrings.Localizable.buttonLogin, isLoading: isLoading, action: login)
                     Button {
@@ -146,12 +184,12 @@ struct OnboardingView: View {
             do {
                 _ = try await AccountManager.instance.createAndSetCurrentAccount(code: code, codeVerifier: verifier)
                 MatomoUtils.connectUser()
-                (self.window?.windowScene?.delegate as? SceneDelegate)?.showMainView()
+                await (self.window?.windowScene?.delegate as? SceneDelegate)?.showMainView()
             } catch {
                 if let previousAccount = previousAccount {
                     AccountManager.instance.switchAccount(newAccount: previousAccount)
                 }
-                IKSnackBar.showSnackBar(message: error.localizedDescription)
+                await IKSnackBar.showSnackBar(message: error.localizedDescription)
             }
             isLoading = false
         }
